+++ package.json
@@ -80,6 +80,7 @@
     "@types/react-autosuggest": "^10.1.5",
     "@types/react-dom": "^17.0.8",
     "@types/react-modal": "^3.13.1",
+    "@types/sanitize-html": "^2.8.0",
     "@types/uuid": "^8.3.0",
     "@types/webpack-env": "^1.16.0",
     "@types/ws": "^7.4.5",
